metal: attention wsum re-grid, f16 W_k, swiglu merge -- best 44.4 tok/s (#86)
Three squeezes from the phase ledger.

The attention kernel kept one threadgroup per head for its weighted sum
-- the bandwidth half, O(seq * kvr) -- which is tolerable at seq 40 and
collapses at 4,096. Split: scores and softmax stay per head and emit
probabilities; a new wsum kernel takes one SIMD group per output
element, the same shape the absorb fix used.

W_k is stored f16, halving the largest reducible read outside the
experts (~113 -> ~56 MB/token). Rounding a weight already quantized to
~4.5 bits is noise against the quantization, and the f64 oracle passes
unchanged.

The six per-slot SwiGLU dispatches are one dispatch (slot = gid / n),
and the barrier between the routed reduce and the shared expert's
gate/up is gone -- they touch disjoint slots.

Best run 44.4 tok/s, band 39.7-44.4 against the prior 41.4-42.9; at
this context length the gains are ~1 ms effects, and the wsum re-grid
is the one that matters as context grows.

The attention split reproduced the canonical bug an eighth time: the
layer tail's own attention dispatch -- a third call site, again -- kept
the fused semantics and fed probabilities to W_v as o_latent. The
layer-tail differential caught it before any manual run.

+++ src/shaders/metal/mla_absorb.metal
@@ -6,11 +6,11 @@
 // and axpy'd it, which forced a synchronization in the middle of every layer
 // whatever else was recorded around it.
 //
-// `W_k` arrives pre-dequantized to f32 and device-resident. That is a
-// deliberate trade: `kv_b` is 8 MB a layer against ~45 MB of experts, it is
-// read on every layer of every token, and holding it as f32 turns this from a
-// per-element K-quant unpack -- which needs block lookup per output column --
-// into a plain transposed matvec.
+// `W_k` arrives pre-dequantized to f16 and device-resident: the same trade as
+// f32 -- no per-element K-quant unpack in the transposed access -- at half the
+// bandwidth, which at ~113 MB per token as f32 was the largest single
+// reducible read outside the experts themselves. f16 rounding of a weight
+// already quantized to ~4.5 bits is noise against the quantization itself.
 //
 // Transposed because the tensor is row-major over kvr: row r of W_k is
 // contiguous, and the output is a sum over rows weighted by q_nope[r]. So each
@@ -40,7 +40,7 @@ struct AbsorbDims {
 // terms; gridding one SIMD group per output makes it 8,192 groups on the real
 // model, and the column reads within a group are contiguous across lanes.
 kernel void mla_absorb(
-    device const float *wk  [[buffer(0)]], // [n_heads * stride][kvr], f32
+    device const half  *wk  [[buffer(0)]], // [n_heads * stride][kvr], f16
     device const float *q   [[buffer(1)]],
     device float       *out [[buffer(2)]], // [n_heads][kvr]
     constant AbsorbDims &d  [[buffer(3)]],
@@ -54,11 +54,11 @@ kernel void mla_absorb(
     if (h >= d.n_heads) return;
     const uint i = g - h * d.kvr;
 
-    device const float *wh = wk + (ulong)h * d.stride * d.kvr;
+    device const half *wh = wk + (ulong)h * d.stride * d.kvr;
     device const float *qh = q + (ulong)h * d.q_stride + d.q_off;
 
     float acc = 0.0f;
-    for (uint r = lane; r < d.nope; r += 32) acc += qh[r] * wh[(ulong)r * d.kvr + i];
+    for (uint r = lane; r < d.nope; r += 32) acc += qh[r] * (float)wh[(ulong)r * d.kvr + i];
     acc = simd_sum(acc);
     if (lane == 0) out[(ulong)h * d.kvr + i] = acc;
 }
